Refactoring turkish translation (into new i18n.tr package)
