+++ src/main/scala/net/kogics/kojo/lite/i18n/trInit.scala
@@ -20,92 +20,78 @@
 
 package net.kogics.kojo.lite.i18n
 
-import net.kogics.kojo.lite.CoreBuiltins
-import net.kogics.kojo.lite.Builtins
-import net.kogics.kojo.picture
+import net.kogics.kojo.lite.{CoreBuiltins, Builtins}
 import net.kogics.kojo.xscala.RepeatCommands
-import io.github.jdiemke.triangulation.Triangle2D
+import net.kogics.kojo.core.Turtle
+import java.awt.Color
 
 object TurkishAPI {
-  // some type aliases in Turkish -- Ctrl-t to return type info will also be in turkish
-  type Nesne = Object
-  type Birim = Unit
-  type Her   = Any
-  type Hiç   = Nothing
+  var builtins: CoreBuiltins = _ //unstable reference to module
+  lazy val richBuiltins = builtins.asInstanceOf[Builtins]
 
-  import net.kogics.kojo.core.Turtle
-  import java.awt.{Color, Paint}
-  type Renk = Color
-  type Boya = Paint
+  import net.kogics.kojo.lite.i18n.tr // todo: better here than at the top scope?
+  // Imports are intransitive (not visibly transitive). So, we "export" the ones that should be on the interface
+  // todo: metaprog on these?
+  type Nesne=tr.Nesne
+  type Birim=tr.Birim
+  type Her=tr.Her
+  type HerDeğer=tr.HerDeğer
+  type HerGönder=tr.HerGönder
+  type Yok=tr.Yok
+  type Hiç=tr.Hiç
+  type Renk=tr.Renk
+  type Boya=tr.Boya
+  type Hız=tr.Hız
+  type Nokta=tr.Nokta
+  type Dikdörtgen=tr.Dikdörtgen
+  type Üçgen=tr.Üçgen
+  type İkil=tr.İkil
+  type Seçim=tr.Seçim
+  type Lokma=tr.Lokma
+  type Kısa=tr.Kısa
+  type Sayı=tr.Sayı
+  type Uzun=tr.Uzun
+  type İriSayı=tr.İriSayı
+  type UfakKesir=tr.UfakKesir
+  type Kesir=tr.Kesir
+  type İriKesir=tr.İriKesir
+  type Harf=tr.Harf
+  type Yazı=tr.Yazı
 
-  import net.kogics.kojo.core.{Speed, Slow, Medium, Fast, SuperFast, Point}
-  type Hız  = Speed
-  type Nokta = Point
-  object Nokta {
-    def apply(x: Kesir, y: Kesir) = new Point(x, y)
-    def unapply(p: Nokta) = Some((p.x, p.y))
-  }
-  type Üçgen = Triangle2D
-
-  // Ref: https://docs.scala-lang.org/overviews/scala-book/built-in-types.html
-  type İkil = Boolean
-  type Seçim = Boolean
-  val (doğru, yanlış) = (true, false)
-  // Bit? Parçacık olsun mu adı?
-
-  // We have Byte/Short/Int/Long which all default to Int and BigInt
-  // val n = 1
-  // Sayma sayıları
-  type Lokma   = Byte
-  type Kısa    = Short
-  type Sayı    = Int
-  type Uzun    = Long
-  type İriSayı = BigInt
-  // We have Float/Double which default to Double and BigDecimal
-  //   val x = 1.0
-  // Kesirli sayılar
-  type UfakKesir = Float
-  type Kesir = Double
-  type İriKesir = BigDecimal
-  // Yazı
-  type Harf = Char
-  type Yazı = String
-
-  type Dizi[B] = collection.Seq[B]
-  object Dizi {
-    def apply[B](elems: B*): Seq[B] = Seq.from(elems)
-  }
-  type Dizin[A] = List[A]
-  object Dizin {
-    def apply[A](elems: A*): List[A] = List.from(elems)
-  }
-  val Boş = scala.collection.immutable.Nil
-  type MiskinDizin[C] = LazyList[C]
-  object MiskinDizin {
-    def sayalım(başlangıç: Sayı, kaçarKaçar: Sayı = 1) = LazyList.from(başlangıç, kaçarKaçar)
-  }
-
-  // Used in Conway's game of life code in the tutorial
-  type Sayılar = Vector[Sayı]
-  object Sayılar {
-    def apply(elemanlar: Sayı*): Sayılar = Vector.from(elemanlar)
-  }
-
-  // örnek: if (tanımlıMı(resim.çarpışma(Resim.tuvalınSınırları))) {...} else {...}
+  type Belki[T]=tr.Belki[T]
+  type Biri[T]=tr.Biri[T]
+  val Hiçbiri=tr.Hiçbiri
+  val Biri=tr.Biri
+  // val varMı = tr.varMı _
+  // örnek: if (varMı(resim.çarpışma(Resim.tuvalınSınırları))) {...} else {...}
   // .isDefined yerine
-  def tanımlıMı[T](o: Option[T]): İkil = o match {
+  def varMı[T](o: Belki[T]): İkil = o match {
+    case None    => yanlış
     case Some(x) => doğru
-    case None => yanlış
   }
+  def yokMu[T](o: Belki[T]): İkil = !varMı(o)
 
-  // 
-  val (yavaş, orta, hızlı, çokHızlı) = (Slow, Medium, Fast, SuperFast)
+  type Dizi[T]=tr.Dizi[T]
+  type Dizin[T]=tr.Dizin[T]
+  type MiskinDizin[T]=tr.MiskinDizin[T]
 
-  var builtins: net.kogics.kojo.lite.CoreBuiltins = _ //unstable reference to module
+  type Sayılar=tr.Sayılar
+  type UzunlukBirimi=tr.UzunlukBirimi
 
-  import net.kogics.kojo.core
-  type UzunlukBirimi = core.UnitLen
-  val (noktaSayısı, santim, inç) = (core.Pixel, core.Cm, core.Inch)
+  type GeoYol=tr.GeoYol
+  type GeoNokta=tr.GeoNokta
+  type Grafik2B=tr.Grafik2B
+
+  val (doğru, yanlış, yavaş, orta, hızlı, çokHızlı, noktaSayısı, santim, inç) = (tr.doğru, tr.yanlış, tr.yavaş, tr.orta, tr.hızlı, tr.çokHızlı, tr.noktaSayısı, tr.santim, tr.inç)
+
+  val Nokta = tr.Nokta
+  val Dizi = tr.Dizi
+  val Dizin = tr.Dizin
+  val Sayılar = tr.Sayılar
+  val MiskinDizin = tr.MiskinDizin
+  val Boş = tr.Boş
+
+  def gerekli(gerekçe: İkil, mesaj: => Any): Birim = require(gerekçe, mesaj)
 
   trait TurkishTurtle {
     def englishTurtle: Turtle
@@ -278,27 +264,11 @@ object TurkishAPI {
   def satıryaz() = println()
   def yaz(data: Her) = print(data)
 
-  //math functions
-  def yuvarla(sayı: Number, basamaklar: Sayı = 0): Kesir = {
-    val faktor = math.pow(10, basamaklar).toDouble
-    math.round(sayı.doubleValue * faktor).toLong / faktor
-  }
-  def piSayısı(): Kesir = math.Pi
-  def eSayısı(): Kesir = math.E
-  def mutlakDeğer(x: Kesir): Kesir = math.abs(x)
-  def karesi(x: Kesir): Kesir = math.pow(x, 2)
-  def karekökü(x: Kesir): Kesir = math.sqrt(x)
-  val gücü = kuvveti _
-  def kuvveti(x: Kesir, k: Kesir): Kesir = math.pow(x, k)
-  def onlukTabandaLogu(x: Kesir): Kesir = math.log10(x)
-  def doğalLogu(x: Kesir): Kesir = math.log(x)
-  def logaritması(x: Kesir): Kesir = math.log(x)
-  def sinüs(x: Kesir): Kesir = math.sin(x)
-  def kosinüs(x: Kesir): Kesir = math.cos(x)
-  def tanjant(x: Kesir): Kesir = math.tan(x)
-  def sinüsünAçısı(x: Kesir): Kesir = math.asin(x)
-  def kosinüsünAçısı(x: Kesir): Kesir = math.acos(x)
-  def tanjantınAçısı(x: Kesir): Kesir = math.atan(x)
+  import tr.{matematik => m}
+  def piSayısı: Kesir = math.Pi
+  def eSayısı: Kesir = math.E
+  val gücü = m.kuvveti _
+  val (yuvarla, mutlakDeğer, karesi, karekökü, kuvveti, onlukTabandaLogu, doğalLogu, logaritması, sinüs, kosinüs, tanjant, sinüsünAçısı, kosinüsünAçısı, tanjantınAçısı) = (m.yuvarla _, m.mutlakDeğer _, m.karesi _, m.karekökü _, m.kuvveti _, m.onlukTabandaLogu _, m.doğalLogu _, m.logaritması _, m.sinüs _, m.kosinüs _, m.tanjant _, m.sinüsünAçısı _, m.kosinüsünAçısı _, m.tanjantınAçısı _)
   // todo: don't we have a type class for Num?
   def enİrisi(x: Sayı, y: Sayı): Sayı = math.max(x, y)
   def enUfağı(x: Sayı, y: Sayı): Sayı = math.min(x, y)
@@ -308,7 +278,7 @@ object TurkishAPI {
   def enUfağı(x: Kesir, y: Kesir ): Kesir = math.min(x, y)
   def enİrisi(x: UfakKesir, y: UfakKesir ): UfakKesir = math.max(x, y)
   def enUfağı(x: UfakKesir, y: UfakKesir ): UfakKesir = math.min(x, y)
-  // todo more...
+
   // ../CoreBuiltins.scala
   def rastgele() = math.random()
   def rastgele(üstSınır: Sayı) = builtins.random(üstSınır)
@@ -327,6 +297,9 @@ object TurkishAPI {
   def rastgeleŞeffafRenk = builtins.randomTransparentColor
   def rastgeleDiziden[T](dizi: Dizi[T]) = builtins.randomFrom(dizi)
   def rastgeleDiziden[T](dizi: Dizi[T], ağırlıklar: Dizi[Kesir]) = builtins.randomFrom(dizi, ağırlıklar)
+  // def diziKarıştır[T](xs: Dizi[T]): Dizi[T] = util.Random.shuffle(xs)
+  def rastgeleKarıştır[T, C](xs: IterableOnce[T])(implicit bf: collection.BuildFrom[xs.type, T, C]): C = util.Random.shuffle(xs)
+
   def durakla(saniye: Kesir) = builtins.pause(saniye)
 
   def üçgenDöşeme(noktalar: Dizi[Nokta]): Dizi[Üçgen] = builtins.triangulate(noktalar)
@@ -335,49 +308,13 @@ object TurkishAPI {
   def buDizin = builtins.currentDir
   def kurulumDizini = builtins.installDir
   def yazıyüzleri = builtins.availableFontNames
+  def yazıyüzü(adı: Yazı, boyu: Sayı) = builtins.Font(adı, boyu)
+  def yazıyüzü(adı: Yazı, boyu: Sayı, biçem: Sayı) = builtins.Font(adı, biçem, boyu)
+  def yazıÇerçevesi(yazı: Yazı, yazıBoyu: Sayı, yazıAdı: Yazı = null): Dikdörtgen = builtins.textExtent(yazı, yazıBoyu, yazıAdı)
 
   val kaplumbağa0 = kaplumbağa
   def yeniKaplumbağa(x: Kesir, y: Kesir) = new Kaplumbağa(x, y)
   def yeniKaplumbağa(x: Kesir, y: Kesir, giysiDosyası: Yazı) = new Kaplumbağa(x, y, giysiDosyası)
-
-  lazy val richBuiltins = builtins.asInstanceOf[Builtins]
-
-  type ResimDosyası = richBuiltins.Image
-  // ../../util/Vector2D.scala
-  case class Yöney2B(x: Kesir, y: Kesir) {
-    def this(v: richBuiltins.Vector2D) = this(v.x, v.y)
-    val v = richBuiltins.Vector2D(x, y)
-    def döndür(açı: Kesir) = Yöney2B(v.rotate(açı))
-    def büyüt(oran: Kesir) = Yöney2B(v.scale(oran))
-    def boyunuBirYap = Yöney2B(v.normalize)
-    def boyu = v.magnitude
-    def boyunKaresi = v.magSquared
-    def sınırla(sınır: Kesir): Yöney2B = {
-      if (boyu < sınır) this
-      else boyunuBirYap * sınır
-    }
-    def +(y2: Yöney2B): Yöney2B = Yöney2B(v + y2.v)
-    def -(y2: Yöney2B): Yöney2B = Yöney2B(v - y2.v)
-    def *(oran: Kesir): Yöney2B = Yöney2B(v * oran)
-    def /(oran: Kesir): Yöney2B = Yöney2B(v / oran)
-    def içÇarpım(y2: Yöney2B): Kesir = v.dot(y2.v)
-    def izdüşümü(y2: Yöney2B): Yöney2B = Yöney2B(v.project(y2.v))
-    def ağırlıklıToplam(y2: Yöney2B, oran: Kesir): Yöney2B = Yöney2B(v.lerp(y2.v, oran)) // lerp == weightedSum
-    def uzaklık(y2: Yöney2B): Kesir = v.distance(y2.v)
-    def doğrultu = v.heading
-    def açı(y2: Yöney2B): Kesir = v.angle(y2.v)
-    def açı2(y2: Yöney2B): Kesir = v.angleTo(y2.v)
-    def unary_- : Yöney2B = Yöney2B(-v)
-    def yansıt(y2: Yöney2B): Yöney2B = Yöney2B(v.bounceOff(y2.v))
-    override def toString = "Yöney2B(%.2f, %.2f)".format(x, y)
-    override def equals(y2: Any) = y2 match {
-      case y3: Yöney2B => v.equals(y3.v)
-      case _ => yanlış
-    }
-  }
-  object Yöney2B {
-    def apply(v: richBuiltins.Vector2D) = new Yöney2B(v.x, v.y)
-  }
 
   def buradaDur = burdaDur _
   def burdaDur(mesaj: Any) = richBuiltins.breakpoint(mesaj)
@@ -423,292 +360,6 @@ object TurkishAPI {
   def yatayMerkezKonumu(uzunluk: Kesir): Kesir = tuvalAlanı.x + (tuvalAlanı.en - uzunluk) / 2
   def dikeyMerkezKonumu(uzunluk: Kesir): Kesir = tuvalAlanı.y + (tuvalAlanı.boy - uzunluk) / 2
 
-  // ../../picture/transforms.scala
-  abstract class ComposableTransformer extends Function1[Resim, Resim] { outer =>
-    def apply(r: Resim): Resim
-    def -> (r: Resim) = apply(r)
-    def *(digeri: ComposableTransformer) = new ComposableTransformer {
-      def apply(r: Resim): Resim = outer.apply(digeri.apply(r))
-    }
-  }
-  // ../../picture/picimage.scala
-  abstract class ComposableImageEffect extends ComposableTransformer {
-    def epic(r: Resim) = r.p match {
-      case ep: picture.EffectablePicture => ep
-      case _                             => new picture.EffectableImagePic(r.p)(r.p.canvas)
-    }
-  }
-
-  case class Rotc(angle: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Rot(angle)(r.p)) }
-  case class Rotpc(angle: Double, x: Double, y: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Rotp(angle, x, y)(r.p)) }
-  case class Scalec(factor: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Scale(factor)(r.p)) }
-  case class ScaleXYc(x: Double, y: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.ScaleXY(x, y)(r.p)) }
-  case class Opacc(f: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Opac(f)(r.p)) }
-  case class Huec(f: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Hue(f)(r.p)) }
-  case class Satc(f: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Sat(f)(r.p)) }
-  case class Britc(f: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Brit(f)(r.p)) }
-  case class Transc(x: Double, y: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Trans(x, y)(r.p)) }
-  case class Offsetc(x: Double, y: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Offset(x, y)(r.p)) }
-  case object FlipYc extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.FlipY(r.p)) }
-  case object FlipXc extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.FlipX(r.p)) }
-  case object AxesOnc extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.AxesOn(r.p)) }
-  case class Fillc(color: Paint) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Fill(color)(r.p)) }
-  case class Strokec(color: Paint) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Stroke(color)(r.p)) }
-  case class StrokeWidthc(w: Double) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.StrokeWidth(w)(r.p)) }
-  case class PreDrawTransformc(fn: Resim => Birim) extends ComposableTransformer {
-    def apply(r: Resim) = {
-      val f2 = new Function1[richBuiltins.Picture, Unit] {
-        def apply(p: richBuiltins.Picture): Unit = fn(r)
-      }
-      new Resim(picture.PreDrawTransform(f2)(r.p))
-    }
-  }
-  case class PostDrawTransformc(fn: Resim => Birim) extends ComposableTransformer {
-    def apply(r: Resim) = {
-      val f2 = new Function[richBuiltins.Picture, Unit] {
-        def apply(p: richBuiltins.Picture): Unit = fn(r)
-      }
-      new Resim(picture.PostDrawTransform(f2)(r.p))
-    }
-  }
-  // ../../picture/effects.scala
-  case class Spinc(n: Int) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Spin(n)(r.p)) }
-  case class Reflectc(n: Int) extends ComposableTransformer { def apply(r: Resim) = new Resim(picture.Reflect(n)(r.p)) }
-  // ../../picture/picimage.scala
-  case class Fadec(n: Int) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.Fade(n)(epic(r)))
-  }
-  case class Blurc(n: Int) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.Blur(n)(epic(r)))
-  }
-  case class PointLightc(x: Double, y: Double, direction: Double, elevation: Double, distance: Double) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.PointLightEffect(x, y, direction, elevation, distance)(epic(r)))
-  }
-  case class SpotLightc(x: Double, y: Double, direction: Double, elevation: Double, distance: Double) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.SpotLightEffect(x, y, direction, elevation, distance)(epic(r)))
-  }
-  case class Lightsc(lights: com.jhlabs.image.LightFilter.Light*) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.Lights(lights: _*)(epic(r)))
-  }
-  case class Noisec(amount: Int, density: Double) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.Noise(amount, density)(epic(r)))
-  }
-  case class Weavec(xWidth: Double, xGap: Double, yWidth: Double, yGap: Double) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.Weave(xWidth, xGap, yWidth, yGap)(epic(r)))
-  }
-  case class SomeEffectc(name: Symbol, props: Tuple2[Symbol, Any]*) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.SomeEffect(name, props: _*)(epic(r)))
-  }
-  case class ApplyFilterc(filter: java.awt.image.BufferedImageOp) extends ComposableImageEffect {
-    def apply(r: Resim) = new Resim(picture.ApplyFilter(filter)(epic(r)))
-  }
-  // ../../picture/package.scala
-  def döndür(açı: Kesir) = Rotc(açı)
-  def döndürMerkezli(açı: Kesir, x: Kesir, y: Kesir) = Rotpc(açı, x, y)
-  def büyüt(oran: Kesir) = Scalec(oran)
-  def büyüt(xOranı: Kesir, yOranı: Kesir) = büyütXY(xOranı, yOranı)
-  def büyütXY(xOranı: Kesir, yOranı: Kesir) = ScaleXYc(xOranı, yOranı)
-  def saydamlık(oran: Kesir) = Opacc(oran)
-  def ton(t: Kesir) = Huec(t)
-  def parlaklık(p: Kesir) = Satc(p)
-  def aydınlık(a: Kesir) = Britc(a)
-  def götür(x: Kesir, y: Kesir) = Transc(x, y)
-  def kaydır(x: Kesir, y: Kesir) = Offsetc(x, y)
-  def yansıtY = FlipYc
-  def yansıtX = FlipXc
-  def eksenler = AxesOnc
-  def boyaRengi(r: Boya) = Fillc(r)
-  def kalemRengi(r: Boya) = Strokec(r)
-  def kalemBoyu(b: Kesir) = StrokeWidthc(b)
-  def çizimÖncesiİşlev(iv: Resim => Birim) = PreDrawTransformc(iv)
-  def çizimSonrasıİşlev(iv: Resim => Birim) = PostDrawTransformc(iv)
-  def çevir(sayı: Sayı) = Spinc(sayı)
-  def yansıt(sayı: Sayı) = Reflectc(sayı)
-  def soluk(n: Sayı) = Fadec(n)
-  def bulanık(n: Sayı) = Blurc(n)
-  def noktaIşık(x: Kesir, y: Kesir, yön: Kesir, yükseklik: Kesir, uzaklık: Kesir) = PointLightc(x, y, yön, yükseklik, uzaklık)
-  def sahneIşığı(x: Kesir, y: Kesir, yön: Kesir, yükseklik: Kesir, uzaklık: Kesir) = SpotLightc(x, y, yön, yükseklik, uzaklık)
-  def ışıklar(ışıklar: com.jhlabs.image.LightFilter.Light*) = Lightsc(ışıklar: _*)
-  def birEfekt(isim: Symbol, özellikler: Tuple2[Symbol, Any]*) = SomeEffectc(isim, özellikler: _*)
-  def filtre(filtre: java.awt.image.BufferedImageOp) = ApplyFilterc(filtre)
-  def gürültü(miktar: Sayı, yoğunluk: Kesir) = Noisec(miktar, yoğunluk)
-  def örgü(xBoyu: Kesir, xAra: Kesir, yBoyu: Kesir, yAra: Kesir) = Weavec(xBoyu, xAra, yBoyu, yAra)
-  def NoktaIşık(x: Kesir, y: Kesir, yön: Kesir, yükseklik: Kesir, uzaklık: Kesir) = picture.PointLight(x, y, yön, yükseklik, uzaklık)
-  def SahneIşığı(x: Kesir, y: Kesir, yön: Kesir, yükseklik: Kesir, uzaklık: Kesir) = picture.SpotLight(x, y, yön, yükseklik, uzaklık)
-  // ../../core/Picture.scala
-
-  // ../../core/vertexShapeSupport.scala
-  type GeoYol = java.awt.geom.GeneralPath
-  type GeoNokta = core.VertexShape
-  type Grafik2B = scala.swing.Graphics2D
-
-  class Resim(val p: richBuiltins.Picture) {
-    def tuval = p.canvas
-    def pnode = p.pnode // ??
-    def tnode = p.tnode // ??
-    def çiz() = p.draw()
-    def çizili: İkil = p.isDrawn
-    def sınırlar = p.bounds
-    def döndür(açı: Kesir): Birim = p.rotate(açı)
-    def döndürMerkezli(açı: Kesir, x: Kesir, y: Kesir): Birim = p.rotateAboutPoint(açı, x, y)
-    def büyüt(oran: Kesir) = p.scale(oran)
-    def büyüt(x: Kesir, y: Kesir) = p.scale(x, y)
-    def götür(x: Kesir, y: Kesir) = p.translate(x, y)
-    def götür(yy: Yöney2B) = p.translate(yy.v.x, yy.v.y)
-    def hızınıDönüştür(yy: Yöney2B) = p.transv(yy.v)
-    def kaydır(x: Kesir, y: Kesir) = p.offset(x, y)
-    def kaydır(yy: Yöney2B) = p.offset(yy.v.x, yy.v.y)
-    def yansıtX() = p.flipX()
-    def yansıtY() = p.flipY()
-    def saydamlık(oran: Kesir) = p.opacityMod(oran)
-    def ton(t: Kesir) = p.hueMod(t)
-    def parlaklık(f: Kesir) = p.satMod(f)
-    def aydınlık(f: Kesir) = p.britMod(f)
-    def benzerDönüşümUygula(bd: java.awt.geom.AffineTransform) = p.transformBy(bd)
-    def benzerDönüşümKur(bd: java.awt.geom.AffineTransform) = p.setTransform(bd)
-    def bilgiVer() = p.dumpInfo()
-    def kopya: Resim = new Resim(p.copy)
-    def eksenleriGöster() = p.axesOn()
-    def eksenleriGizle() = p.axesOff()
-    def sil() = p.erase()
-    def göster() = p.visible()
-    def gizle() = p.invisible()
-    def görünürlüğüTersineÇevir() = p.toggleV()
-    def görünür: İkil = p.isVisible
-    def kesişir(başkaResim: Resim): İkil = p.intersects(başkaResim.p)
-    def çarptıMı = çarpıştı _
-    def çarpıştı(başkaResim: Resim): İkil = p.intersects(başkaResim.p)
-    def çarpışmalar(başkaları: Set[Resim]): Set[Resim] = 
-      başkaları.filter {this çarpıştı _}
-    def çarpışma(başkaları: Seq[Resim]): Option[Resim] =
-      başkaları.find {this çarpıştı _}
-    def kesişim(başkaResim: Resim): com.vividsolutions.jts.geom.Geometry =
-      p.intersection(başkaResim.p)
-    def içinde(başkaResim: Resim) = p.contains(başkaResim.p)
-    def uzaklık(başkaResim: Resim) = p.distanceTo(başkaResim.p)
-    def alan = p.area
-    def çevre = p.perimeter
-    def geometri = p.picGeom
-    def konum = p.position
-    val konumuKur = kondur _
-    def kondur(x: Kesir, y: Kesir) = p.setPosition(x, y)
-    def doğrultu = p.heading
-    val doğrultuyuKur = açıyaDön _
-    def açıyaDön(açı: Kesir) = p.setHeading(açı)
-    def büyütmeOranı: (Kesir, Kesir) = p.scaleFactor
-    def büyütmeOranınıKur(x: Kesir, y: Kesir) = p.setScaleFactor(x, y)
-    def büyütmeyiKur(oran: Kesir) = p.setScale(oran)
-    def dönüşüm = p.transform
-    def kalemRenginiKur = p.setPenColor _
-    def kalemKalınlığınıKur(kalınlık: Sayı) = p.setPenThickness(kalınlık)
-    def kalemiKapa() = p.setNoPen()
-    //def setPenCapJoin(capJoin: (Int, Int)): Unit = setPenCapJoin(capJoin._1, capJoin._2)
-    //def setPenCapJoin(cap: Int, join: Int): Unit
-    def boyamaRenginiKur(renk: Boya) = p.setFillColor(renk)
-    def saydamlık = p.opacity
-    def saydamlığıKur(s: Kesir) = p.setOpacity(s)
-    def canlan = tepkiVer _
-    def tepkiVer(fn: Resim => Birim) = {
-      val bu = this
-      p.react(new Function1[richBuiltins.Picture, Unit] {
-        def apply(p: richBuiltins.Picture) = fn(bu)
-      })
-
-    }
-    def tepkileriDurdur() = p.stopReactions()
-    def dönüştür(fn: Seq[richBuiltins.PolyLine] => Seq[richBuiltins.PolyLine]) = p.morph(fn)
-    def yinele(fn: richBuiltins.PolyLine => Unit) = p.foreachPolyLine(fn)
-    // ../../picture/pics.scala
-    def boşluk(uzunluk: Kesir): Resim = p match {
-      case rd: picture.BasePicList => new Resim(rd.withGap(uzunluk))
-      case _ => this
-    }
-    def imge = p.toImage
-    def girdiyiAktar(r: Resim) = p.forwardInputTo(r.p)
-    def öneAl() = p.moveToFront()
-    def ardaAl() = p.moveToBack()
-    def arkayaAl() = ardaAl()
-    def sonrakiniGöster(ara: Uzun = 100) = p.showNext(ara)
-    def güncelle(yeniVeri: Her) = p.update(yeniVeri)
-    def çizimiKontrolEt(mesaj: Yazı) = p.checkDraw(mesaj)
-    def yana(başka: Resim) = p.beside(başka.p)
-    def üste(başka: Resim) = p.above(başka.p)
-    def alta(başka: Resim) = p.below(başka.p)
-    def konumaDoğruGit(x: Kesir, y: Kesir, süre: Uzun)(sonİşlev: => Birim) =
-      p.animateToPosition(x, y, süre)(sonİşlev)
-    def mesafeGit(mx: Kesir, my: Kesir, süre: Uzun)(sonİşlev: => Birim) =
-      p.animateToPositionDelta(mx, my, süre)(sonİşlev)
-    def fareyeTıklayınca(iş: (Kesir, Kesir) => Birim) = p.onMouseClick(iş)
-    def fareyiSürükleyince(iş: (Kesir, Kesir) => Birim) = p.onMouseDrag(iş)
-    def fareKımıldayınca(iş: (Kesir, Kesir) => Birim) = p.onMouseMove(iş)
-    def fareyeBasınca(iş: (Kesir, Kesir) => Birim) = p.onMousePress(iş)
-    def fareyiBırakınca(iş: (Kesir, Kesir) => Birim) = p.onMouseRelease(iş)
-    def fareGirince(iş: (Kesir, Kesir) => Birim) = p.onMouseEnter(iş)
-    def fareÇıkınca(iş: (Kesir, Kesir) => Birim) = p.onMouseExit(iş)
-    // todo: more? see: https://docs.kogics.net/reference/picture.html
-  }
-  // ../../picture/package.scala 
-  // ../Builtins.scala  object Picture
-  object Resim {
-    def apply(işlev: => Birim): Resim = new Resim(richBuiltins.Picture(işlev))
-    def çiz(r: Resim) = richBuiltins.draw(r.p)
-    val doğru = köşegen _
-    def köşegen(en: Kesir, boy: Kesir) = new Resim(richBuiltins.Picture.line(en, boy))
-    def yay(yarıçap: Kesir, açı: Kesir) = new Resim(richBuiltins.Picture.arc(yarıçap, açı))
-    def daire(yarıçap: Kesir) = new Resim(richBuiltins.Picture.circle(yarıçap))
-    def elips(xYarıçapı: Kesir, yYarıçapı: Kesir) = new Resim(richBuiltins.Picture.ellipse(xYarıçapı, yYarıçapı))
-    def elipsDikdörtgenİçinde(en: Kesir, boy: Kesir) = new Resim(richBuiltins.Picture.ellipseInRect(en, boy))
-    def yatay(boy: Kesir) = new Resim(richBuiltins.Picture.hline(boy))
-    def dikey(boy: Kesir) = new Resim(richBuiltins.Picture.vline(boy))
-    def dikdörtgen(en: Kesir, boy: Kesir) = new Resim(richBuiltins.Picture.rect(boy, en)) // they are swapped!
-    // ../../Picture/package.scala
-    def satır(r: => Resim, kaçTane: Sayı) = new Resim(picture.row(r.p, kaçTane))
-    def sütun(r: => Resim, kaçTane: Sayı) = new Resim(picture.col(r.p, kaçTane))
-    def yazı(içerik: Her, yazıBoyu: Sayı=15) = new Resim(richBuiltins.Picture.text(içerik, yazıBoyu))
-    def yazıRenkli(içerik: Her, yazıBoyu: Sayı, renk: Renk) = new Resim(richBuiltins.Picture.textu(içerik, yazıBoyu, renk))
-    def imge(dosyaAdı: Yazı) = new Resim(richBuiltins.Picture.image(dosyaAdı))
-    def imge(dosyaAdı: Yazı, zarf: Resim) = new Resim(richBuiltins.Picture.image(dosyaAdı, zarf.p))
-    def imge(url: java.net.URL) = new Resim(richBuiltins.Picture.image(url))
-    def imge(url: java.net.URL, zarf: Resim) = new Resim(richBuiltins.Picture.image(url, zarf.p))
-    def imge(imge: ResimDosyası) = new Resim(richBuiltins.Picture.image(imge))
-    def imge(imge: ResimDosyası, zarf: Resim) = new Resim(richBuiltins.Picture.image(imge, zarf.p))
-    // Resim.düğme("Merhaba")(println(kg.x))
-    def düğme(ad: Yazı)(işlev: => Birim) = new Resim(richBuiltins.Picture.button(ad)(işlev))
-    // Resim.arayüz(Label("Merhaba"))
-    // Resim.arayüz(Button("Merhaba")(println("Selam!")))
-    def arayüz(parça: javax.swing.JComponent) = new Resim(richBuiltins.Picture.widget(parça))
-    def yatayBoşluk(en:  Kesir) = new Resim(richBuiltins.Picture.hgap(en))
-    def dikeyBoşluk(boy: Kesir) = new Resim(richBuiltins.Picture.vgap(boy))
-    def yoldan(işlev: GeoYol => Birim) = new Resim(richBuiltins.Picture.fromPath(işlev))
-    def noktadan(işlev: GeoNokta => Birim) = new Resim(richBuiltins.Picture.fromVertexShape(işlev))
-    def kaplumbağadan(işlev: Kaplumbağa => Birim) = { // todo: does this work?
-      val f = new Function1[Turtle, Unit] { def apply(t: Turtle): Unit = işlev(new Kaplumbağa(t)) }
-      new Resim(richBuiltins.Picture.fromTurtle(f))
-    }
-    def tuvalden(en: Kesir, boy: Kesir)(işlev: Grafik2B => Birim) = new Resim(richBuiltins.Picture.fromCanvas(en, boy)(işlev))
-    def eksenleriGöster(r: Resim) = richBuiltins.Picture.showAxes(r.p)
-    def eksenleriGöster(resimler: Resim*) = richBuiltins.Picture.showAxes(resimler.map(_.p): _*)
-    def sınırlarıGöster(r: Resim) = richBuiltins.Picture.showBounds(r.p)
-    def sınırlarıGöster(resimler: Resim*) = richBuiltins.Picture.showBounds(resimler.map(_.p): _*)
-    // todo: more..
-    def tuval = tuvalSınırları  // stageBorder
-    def tuvalinSınırları = tuvalSınırları
-    def tuvalSınırları = new Resim(richBuiltins.tCanvas.stage)
-    def tuvalinSolu = new Resim(richBuiltins.tCanvas.stageLeft)
-    def tuvalinSağı = new Resim(richBuiltins.tCanvas.stageRight)
-    def tuvalinTavanı = new Resim(richBuiltins.tCanvas.stageTop)
-    def tuvalinTabanı = new Resim(richBuiltins.tCanvas.stageBot)
-    def tuvalBölgesi = new Resim(richBuiltins.tCanvas.stageArea)
-  }
-  def çiz(r: Resim) = Resim.çiz(r)
-  def çiz(rler: Resim*) = richBuiltins.draw(rler.map(_.p): _*)
-  def çiz(rler: collection.Seq[Resim]) = richBuiltins.draw(rler.map(_.p))
-  def çizMerkezde(r: Resim) = richBuiltins.drawCentered(r.p)
-  def çizSahne(boya: Paint) = richBuiltins.tCanvas.drawStage(boya)
-  def çizMerkezdeYazı(mesaj: Yazı, renk: Renk, yazıBoyu: Sayı) = richBuiltins.drawCenteredMessage(mesaj, renk, yazıBoyu)
-  def çizVeSakla(resimler: Resim*) = richBuiltins.drawAndHide(resimler.map(_.p): _*)
-  def merkezeTaşı(resim: Resim) = richBuiltins.center(resim.p)
-  def resimleriSil() = richBuiltins.tCanvas.erasePictures()
   def ekranTazelemeHızınıKur(saniyedeKaçKere: Sayı) = richBuiltins.setRefreshRate(saniyedeKaçKere)
   def ekranTazelemeHızınıGöster(renk: Renk, yazıBoyu: Sayı = 15): Birim = { // richBuiltins.showFps(renk, yazıBoyu)
     val cb = richBuiltins.canvasBounds
@@ -726,21 +377,6 @@ object TurkishAPI {
       frameCnt += 1
     }
   }
-
-  def resimDizisi(rd: Resim*) = new Resim(richBuiltins.picStack(rd.map(_.p)))
-  def resimDikeyDizi(rd: Resim*) = new Resim(richBuiltins.picCol(rd.map(_.p)))
-  def resimYatayDizi(rd: Resim*) = new Resim(richBuiltins.picRow(rd.map(_.p)))
-  def resimDüzenliDizi(rd: Resim*) = new Resim(richBuiltins.picStackCentered(rd.map(_.p)))
-  def resimDikeyDüzenliDizi(rd: Resim*) = new Resim(richBuiltins.picColCentered(rd.map(_.p)))
-  def resimYatayDüzenliDizi(rd: Resim*) = new Resim(richBuiltins.picRowCentered(rd.map(_.p)))
-  def resimKümesi(rd: Resim*) = new Resim(richBuiltins.picBatch(rd.map(_.p)))
-  def resimDizisi(rd: collection.Seq[Resim]) = new Resim(richBuiltins.picStack(rd.map(_.p)))
-  def resimDikeyDizi(rd: collection.Seq[Resim]) = new Resim(richBuiltins.picCol(rd.map(_.p)))
-  def resimYatayDizi(rd: collection.Seq[Resim]) = new Resim(richBuiltins.picRow(rd.map(_.p)))
-  def resimDüzenliDizi(rd: collection.Seq[Resim]) = new Resim(richBuiltins.picStackCentered(rd.map(_.p)))
-  def resimDikeyDüzenliDizi(rd: collection.Seq[Resim]) = new Resim(richBuiltins.picColCentered(rd.map(_.p)))
-  def resimYatayDüzenliDizi(rd: collection.Seq[Resim]) = new Resim(richBuiltins.picRowCentered(rd.map(_.p)))
-  def resimKümesi(rd: collection.Seq[Resim]) = new Resim(richBuiltins.picBatch(rd.map(_.p)))
 
   // ../DrawingCanvasAPI.scala
   def yaklaş(oran: Kesir) = richBuiltins.tCanvas.zoom(oran)
@@ -781,20 +417,12 @@ object TurkishAPI {
   def canlandırmaBitince(işlev: => Birim) = richBuiltins.tCanvas.onAnimationStop(işlev)
   def tuvaliEtkinleştir() = richBuiltins.activateCanvas()
   def yazılımcıkDüzenleyicisiniEtkinleştir() = richBuiltins.activateEditor()
-  def sahneKenarındanYansıtma(r: Resim, yöney: Yöney2B): Yöney2B =
-    Yöney2B(richBuiltins.bouncePicOffStage(r.p, yöney.v))
-  def engeldenYansıtma(r: Resim, yöney: Yöney2B, engel: Resim): Yöney2B =
-    Yöney2B(richBuiltins.bouncePicOffPic(r.p, yöney.v, engel.p))
-  def tuşaBasılıMı(tuş: Sayı) = richBuiltins.isKeyPressed(tuş)
 
   def çıktıArtalanınıKur(renk: Renk) = richBuiltins.setOutputBackground(renk)
   def çıktıYazıRenginiKur(renk: Renk) = richBuiltins.setOutputTextColor(renk)
   def çıktıYazıYüzüBoyunuKur(boy: Sayı) = richBuiltins.setOutputTextFontSize(boy)
   def tuvalBoyutOranınınKur(oran: Kesir) = richBuiltins.setDrawingCanvasAspectRatio(oran)
   def tuvalBoyutlarınıKur(en: Sayı, boy: Sayı) = richBuiltins.setDrawingCanvasSize(en, boy)
-
-  def resmiSüz(r: Resim, süzgeç: java.awt.image.BufferedImageOp): Resim = new Resim(richBuiltins.filterPicture(r.p, süzgeç))
-  def resimDosyasınıSüz(rd: java.awt.image.BufferedImage, süzgeç: java.awt.image.BufferedImageOp) = richBuiltins.filterImage(rd, süzgeç)
 
   def süreTut(komut: => Birim): Birim = {
     val t0 = buSaniye
@@ -808,6 +436,72 @@ object TurkishAPI {
 
   def sırayaSok(kaçSaniyeSonra: Kesir)(komut: => Birim) = richBuiltins.schedule(kaçSaniyeSonra)(komut)
   def sırayaSok(n: Sayı, kaçSaniyeSonra: Kesir)(komut: => Birim) = richBuiltins.scheduleN(n, kaçSaniyeSonra)(komut)
+
+  type ResimDosyası = tr.ResimDosyası
+  type Yöney2B = tr.Yöney2B
+  val Yöney2B = tr.Yöney2B
+
+  type Resim = tr.Resim
+  object Resim {
+    import tr.{Resim => R}
+    def apply(işlev: => Birim): Resim = R.apply(işlev)
+    val doğru = R.doğru
+    val (çiz, köşegen, yay, daire, elips, elipsDikdörtgenİçinde, yatay, dikey, dikdörtgen, satır, sütun, yazı, yazıRenkli, düğme, arayüz, yatayBoşluk, dikeyBoşluk, yoldan, noktadan, kaplumbağadan, tuvalden) = (R.çiz _, R.köşegen _, R.yay _, R.daire _, R.elips _, R.elipsDikdörtgenİçinde _, R.yatay _, R.dikey _, R.dikdörtgen _, R.satır _, R.sütun _, R.yazı _, R.yazıRenkli _, R.düğme _, R.arayüz _, R.yatayBoşluk _, R.dikeyBoşluk _, R.yoldan _, R.noktadan _, R.kaplumbağadan _, R.tuvalden _)
+
+    def tuval = tuvalSınırları  // stageBorder
+    def tuvalinSınırları = tuvalSınırları
+    def tuvalSınırları = new Resim(richBuiltins.tCanvas.stage)
+    def tuvalinSolu = new Resim(richBuiltins.tCanvas.stageLeft)
+    def tuvalinSağı = new Resim(richBuiltins.tCanvas.stageRight)
+    def tuvalinTavanı = new Resim(richBuiltins.tCanvas.stageTop)
+    def tuvalinTabanı = new Resim(richBuiltins.tCanvas.stageBot)
+    def tuvalBölgesi = new Resim(richBuiltins.tCanvas.stageArea)
+
+    def imge(dosyaAdı: Yazı) = R.imge(dosyaAdı)
+    def imge(dosyaAdı: Yazı, zarf: Resim) = R.imge(dosyaAdı, zarf)
+    def imge(url: java.net.URL) = R.imge(url)
+    def imge(url: java.net.URL, zarf: Resim) = R.imge(url, zarf)
+    def imge(imge: ResimDosyası) = R.imge(imge)
+    def imge(imge: ResimDosyası, zarf: Resim) = R.imge(imge, zarf)
+
+    def eksenleriGöster(r: Resim) = richBuiltins.Picture.showAxes(r.p)
+    def eksenleriGöster(resimler: Resim*) = richBuiltins.Picture.showAxes(resimler.map(_.p): _*)
+    def sınırlarıGöster(r: Resim) = richBuiltins.Picture.showBounds(r.p)
+    def sınırlarıGöster(resimler: Resim*) = richBuiltins.Picture.showBounds(resimler.map(_.p): _*)
+  }
+  import tr.{res => r}
+  // val döndür = r.döndür _
+  val (döndür, döndürMerkezli) = (r.döndür _, r.döndürMerkezli _)
+  def büyüt(oran: Kesir) = r.büyüt(oran)
+  def büyüt(xOranı: Kesir, yOranı: Kesir) = r.büyüt(xOranı, yOranı)
+  def ışıklar(ışıklar: com.jhlabs.image.LightFilter.Light*) = r.ışıklar(ışıklar: _*)
+  def birEfekt(isim: Symbol, özellikler: Tuple2[Symbol, Any]*) = r.birEfekt(isim, özellikler: _*)
+  val (büyütXY, saydamlık, ton, parlaklık, aydınlık, götür, kaydır, yansıtY, yansıtX, eksenler, boyaRengi, kalemRengi, kalemBoyu, çizimÖncesiİşlev, çizimSonrasıİşlev, çevir, yansıt, soluk, bulanık, noktaIşık, sahneIşığı) = (r.büyütXY _, r.saydamlık _, r.ton _, r.parlaklık _, r.aydınlık _, r.götür _, r.kaydır _, r.yansıtY, r.yansıtX, r.eksenler, r.boyaRengi _, r.kalemRengi _, r.kalemBoyu _, r.çizimÖncesiİşlev _, r.çizimSonrasıİşlev _, r.çevir _, r.yansıt _, r.soluk _, r.bulanık _, r.noktaIşık _, r.sahneIşığı _)
+  val (filtre, gürültü, örgü) = (r.filtre _, r.gürültü _, r.örgü _)
+  def NoktaIşık(x: Kesir, y: Kesir, yön: Kesir, yükseklik: Kesir, uzaklık: Kesir) = r.NoktaIşık(x, y, yön, yükseklik, uzaklık)
+  def SahneIşığı(x: Kesir, y: Kesir, yön: Kesir, yükseklik: Kesir, uzaklık: Kesir) = r.SahneIşığı(x, y, yön, yükseklik, uzaklık)
+  def çiz(r2: Resim) = r.çiz(r2)
+  def çiz(rler: Resim*) = r.çiz(rler: _*)
+  def çiz(rler: collection.Seq[Resim]) = r.çiz(rler)
+  def çizVeSakla(resimler: Resim*) = richBuiltins.drawAndHide(resimler.map(_.p): _*)
+  val (çizMerkezde, çizSahne, çizMerkezdeYazı, merkezeTaşı, resimleriSil) = (r.çizMerkezde _, r.çizSahne _, r.çizMerkezdeYazı _, r.merkezeTaşı _, r.resimleriSil _)
+
+  def resimDizisi(rd: Resim*) = r.resimDizisi(rd)
+  def resimDikeyDizi(rd: Resim*) = r.resimDikeyDizi(rd)
+  def resimYatayDizi(rd: Resim*) = r.resimYatayDizi(rd)
+  def resimDüzenliDizi(rd: Resim*) = r.resimDüzenliDizi(rd)
+  def resimDikeyDüzenliDizi(rd: Resim*) = r.resimDikeyDüzenliDizi(rd)
+  def resimYatayDüzenliDizi(rd: Resim*) = r.resimYatayDüzenliDizi(rd)
+  def resimKümesi(rd: Resim*) = r.resimKümesi(rd)
+  def resimDizisi(rd: collection.Seq[Resim]) = r.resimDizisi(rd)
+  def resimDikeyDizi(rd: collection.Seq[Resim]) = r.resimDikeyDizi(rd)
+  def resimYatayDizi(rd: collection.Seq[Resim]) = r.resimYatayDizi(rd)
+  def resimDüzenliDizi(rd: collection.Seq[Resim]) =r.resimDüzenliDizi(rd)
+  def resimDikeyDüzenliDizi(rd: collection.Seq[Resim]) = r.resimDikeyDüzenliDizi(rd)
+  def resimYatayDüzenliDizi(rd: collection.Seq[Resim]) = r.resimYatayDüzenliDizi(rd)
+  def resimKümesi(rd: collection.Seq[Resim]) = r.resimKümesi(rd)
+
+  val (sahneKenarındanYansıtma, engeldenYansıtma, tuşaBasılıMı, resmiSüz, resimDosyasınıSüz) = (r.sahneKenarındanYansıtma _, r.engeldenYansıtma _, r.tuşaBasılıMı _, r.resmiSüz _, r.resimDosyasınıSüz _)
 
   /* ../../widget/swingwrappers.scala
    Some are used in addition*.scala sample and others:
@@ -827,8 +521,9 @@ object TurkishAPI {
 
 object TurkishInit {
   def init(builtins: CoreBuiltins): Unit = {
-    //initialize unstable value
+    //initialize unstable values:
     TurkishAPI.builtins = builtins
+    tr.builtins = builtins
     builtins match {
       case b: Builtins =>
         println("Kaplumbağalı Kojo'ya Hoşgeldin!")
@@ -1013,3 +708,4 @@ object TurkishInit {
 
   )
 }
+
